backend/middleware_worker/requirements.txt
pika
psycopg2-binary
pymysql
cryptography
